fix(dashboards): add standalone to pipe, move interface to shared
LFXV2-1468
